Add recovery from settings
