+++ packages/browser-wallet/src/popup/pages/Recovery/Recovery.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Route, Routes } from 'react-router-dom';
+import RecoveryIntro from './RecoveryIntro';
+import RecoveryMain from './RecoveryMain';
+
+export default function Recovery() {
+    return (
+        <Routes>
+            <Route index element={<RecoveryIntro />} />
+            <Route path="main" element={<RecoveryMain className="h-full" />} />
+        </Routes>
+    );
+}
