+++ packages/browser-wallet/src/popup/pages/Recovery/i18n/da.ts
@@ -1,0 +1,9 @@
+import type en from './en';
+
+const t: typeof en = {
+    description:
+        'If you are missing some identities or accounts in your wallet, you can try restoring them here.\n\nIf you used your secret recovery phrase to create more identities and accounts in another installation of the Concordium wallet, you can also use this option to add them to this installation.',
+    restore: 'Restore',
+};
+
+export default t;
