+++ packages/browser-wallet/src/popup/pages/Settings/Settings.tsx
@@ -22,6 +22,9 @@ export default function Settings() {
                 <Link className="settings-page__link" to={absoluteRoutes.home.settings.network.path}>
                     {t('network')}
                 </Link>
+                <Link className="settings-page__link" to={absoluteRoutes.home.settings.recovery.path}>
+                    {t('recover')}
+                </Link>
                 <Link className="settings-page__link" to={absoluteRoutes.home.settings.about.path}>
                     {t('about')}
                 </Link>
