+++ packages/browser-wallet/src/popup/pages/Settings/i18n/da.ts
@@ -3,6 +3,7 @@ import type en from './en';
 const t: typeof en = {
     passcode: 'Skift adgangskode',
     network: 'Netværksindstillinger',
+    recover: 'Genskab ID & konti',
     visual: 'Visuelle indstillinger',
     about: 'Om',
 };
