+++ packages/browser-wallet/src/popup/pages/Settings/i18n/en.ts
@@ -1,6 +1,7 @@
 const t = {
     passcode: 'Change passcode',
     network: 'Network settings',
+    recover: 'Restore IDs & accounts',
     visual: 'Visual settings',
     about: 'About',
 };
