+++ packages/browser-wallet/src/popup/pages/Setup/Setup.scss
@@ -110,48 +110,4 @@ $description-margin-top: rem(25px);
             margin-bottom: rem(12px);
         }
     }
-
-    &__recovery {
-        overflow-y: overlay;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-
-        &__results {
-            width: 100%;
-            text-align: left;
-        }
-
-        &__credential {
-            border-top: rem(1px) dotted $color-petroleum;
-
-            p {
-                margin: rem(5px) 0;
-                font-size: rem(8px);
-            }
-        }
-
-        &__button {
-            min-height: rem(30px);
-            margin: auto auto rem(10px);
-        }
-
-        &__identity {
-            margin-top: rem(10px);
-
-            & > p {
-                font-weight: $font-weight-bold;
-                padding: 0;
-            }
-
-            &:not(:last-child) {
-                border-bottom: rem(2px) solid $color-petroleum;
-            }
-        }
-
-        &__description {
-            margin-top: $description-margin-top;
-            margin-bottom: rem(10px);
-        }
-    }
 }
