+++ packages/browser-wallet/src/popup/shell/Routes.tsx
@@ -20,6 +20,7 @@ import AddAccount from '@popup/pages/AddAccount';
 import { IdentityIssuanceEnd, IdentityIssuanceStart } from '@popup/pages/IdentityIssuance';
 import About from '@popup/pages/About';
 import Login from '@popup/pages/Login/Login';
+import Recovery from '@popup/pages/Recovery/Recovery';
 import ChangePasscode from '@popup/pages/ChangePasscode/ChangePasscode';
 
 type PromptKey = keyof Omit<typeof absoluteRoutes['prompt'], 'path'>;
@@ -117,6 +118,7 @@ export default function Routes() {
                     <Route index element={<Settings />} />
                     <Route element={<ChangePasscode />} path={relativeRoutes.home.settings.passcode.path} />
                     <Route element={<NetworkSettings />} path={relativeRoutes.home.settings.network.path} />
+                    <Route element={<Recovery />} path={`${relativeRoutes.home.settings.recovery.path}/*`} />
                     <Route element={<About />} path={relativeRoutes.home.settings.about.path} />
                 </Route>
                 <Route
