+++ packages/browser-wallet/src/popup/shell/i18n/locales/en.ts
@@ -7,6 +7,7 @@ import signMessage from '@popup/pages/SignMessage/i18n/en';
 import connectionRequest from '@popup/pages/ConnectionRequest/i18n/en';
 import settings from '@popup/pages/Settings/i18n/en';
 import networkSettings from '@popup/pages/NetworkSettings/i18n/en';
+import recovery from '@popup/pages/Recovery/i18n/en';
 import about from '@popup/pages/About/i18n/en';
 import addAccount from '@popup/pages/AddAccount/i18n/en';
 import identityIssuance from '@popup/pages/IdentityIssuance/i18n/en';
@@ -24,6 +25,7 @@ const t = {
     connectionRequest,
     settings,
     networkSettings,
+    recovery,
     addAccount,
     identityIssuance,
     about,
