+++ packages/browser-wallet/src/popup/styles/_components.scss
@@ -30,6 +30,7 @@
 @import '../pages/AddAccount/AddAccount';
 @import '../pages/NetworkSettings/NetworkSettings';
 @import '../pages/ChangePasscode/ChangePasscode';
+@import '../pages/Recovery/Recovery';
 
 // Layouts
 @import '../page-layouts/MainLayout';
