+++ packages/browser-wallet/src/popup/styles/elements/_base.scss
@@ -125,3 +125,7 @@ a {
         color: $color-cta;
     }
 }
+
+pre {
+    font-family: $font-family-roboto;
+}
